refactor(app): split drawings/mod.rs into owned siblings
`drawings/mod.rs` owned the object model, the tool port, the tool handle,
the registry and the whole collection with its undo history, at 2,283
production lines. Three siblings now own one responsibility each:

- `tool.rs` — the `DrawingToolImpl` port and the copyable `DrawingTool`
  handle the rest of the app holds a tool by.
- `collection.rs` — what the collection holds, what is selected, hidden or
  locked, and the undo history behind every change.
- `placement.rs` — how a draft is placed anchor by anchor, and how an object
  already on the chart is re-anchored, translated or shifted.

The registry stays in `mod.rs` beside the object model, so a new tool is
still one implementation file plus one name there.

Every body moved unchanged. The only widenings are the `pub(super)` marks
a sibling needs: five methods, six constants, the `DrawingToolImpl` trait
and the `DrawingTool` tuple field. No behaviour, placement, hit-testing or
persisted-format change; no test expectation edited.

`crates/guards/size-baseline.txt`: the `drawings/mod.rs` entry is retired
(865 production lines, under the 1,500 threshold) and `!budget` falls from
45,739 to 43,222.
