+++ crates/app/src/drawings/mod.rs
@@ -18,6 +18,17 @@ mod mark_core;
 mod measure_core;
 mod shape_core;
 
+// The subsystem's three other owners. This module keeps the object model, the
+// vocabulary every tool speaks and the registry; a tool's port and handle live
+// in `tool`, the collection and its undo history in `collection`, and the
+// rules that draft or re-anchor an object in `placement`.
+mod collection;
+mod placement;
+mod tool;
+
+pub use tool::DrawingTool;
+use tool::DrawingToolImpl;
+
 use std::any::Any;
 use std::fmt;
 
@@ -54,20 +65,21 @@ pub const MAX_DRAWING_FILL_ALPHA: u8 = 160;
 /// Undo history depth. One entry per committed command (a whole drag or
 /// slider gesture is one command), so this bounds memory without cutting a
 /// working session short.
-const UNDO_HISTORY_LIMIT: usize = 64;
-const SELECTED_ANCHOR_RADIUS_PX: f32 = 3.5;
+pub(super) const UNDO_HISTORY_LIMIT: usize = 64;
+pub(super) const SELECTED_ANCHOR_RADIUS_PX: f32 = 3.5;
 /// Handles read as hollow rings, not solid discs: the core is the chart's own
 /// backdrop, so the handle marks the anchor without adding a bright blob over
 /// the candles (`docs/ux/drawing-tools-2026-08.md` §D2).
-const SELECTED_ANCHOR_FILL: egui::Color32 = theme::CANVAS;
-const SELECTED_ANCHOR_RING_WIDTH_PX: f32 = 1.25;
+pub(super) const SELECTED_ANCHOR_FILL: egui::Color32 = theme::CANVAS;
+pub(super) const SELECTED_ANCHOR_RING_WIDTH_PX: f32 = 1.25;
 /// Selection never repaints the object white: it keeps the configured colour
 /// and paints this soft halo underneath instead, plus ring anchor handles.
 /// Premultiplied ~11% white — enough to find the object under the pointer,
 /// not enough to double its visual weight.
-const SELECTION_HALO_COLOR: egui::Color32 = egui::Color32::from_rgba_premultiplied(28, 28, 28, 28);
+pub(super) const SELECTION_HALO_COLOR: egui::Color32 =
+    egui::Color32::from_rgba_premultiplied(28, 28, 28, 28);
 /// How much wider than the object's own stroke the halo pass paints.
-const SELECTION_HALO_EXTRA_WIDTH_PX: f32 = 2.5;
+pub(super) const SELECTION_HALO_EXTRA_WIDTH_PX: f32 = 2.5;
 pub(super) const FIB_LABEL_OFFSET_PX: f32 = 3.0;
 pub(super) const FIB_LABEL_SIZE_PX: f32 = 10.0;
 
@@ -432,679 +444,6 @@ pub struct IconLetter {
     pub at: (f32, f32),
     /// Font size as a fraction of the glyph box's height.
     pub height: f32,
-}
-
-/// The implementation port every drawing plugs into. Selection visuals (halo
-/// and anchor handles) are common chrome painted by the wrapper, so a tool
-/// only ever paints its own geometry in the style it is given. Capability
-/// methods drive which inspector sections exist for the tool — an
-/// unsupported property is absent, never disabled.
-trait DrawingToolImpl: Sync {
-    fn id(&self) -> &'static str;
-    /// Human name shown in the inspector header and the object manager.
-    fn name(&self) -> &'static str;
-    fn settings_title(&self) -> &'static str;
-    fn icon(&self) -> &'static str;
-    /// Vector strokes painted in place of [`Self::icon`] when non-empty —
-    /// see [`IconStrokes`].
-    fn icon_strokes(&self) -> IconStrokes {
-        &[]
-    }
-    /// Anchor dots painted over [`Self::icon_strokes`] — see [`IconDots`].
-    /// Only meaningful alongside strokes: a font glyph draws its own.
-    fn icon_dots(&self) -> IconDots {
-        &[]
-    }
-    /// A letter set into the vector icon — see [`IconLetter`]. Only
-    /// meaningful alongside strokes: a font glyph is somebody else's
-    /// drawing, with no corner reserved to put a letter in.
-    fn icon_letter(&self) -> Option<IconLetter> {
-        None
-    }
-    fn hover_text(&self) -> &'static str;
-    fn required_points(&self) -> usize;
-    /// What the *next* click will do, with `placed` anchors already down.
-    ///
-    /// A multi-anchor tool that stops following the pointer looks broken:
-    /// the trader dragged, let go, and the object sat there waiting for a
-    /// click nobody told them about. The rail's `2/3` badge is true but it
-    /// is on the far side of the screen from where the eye is. A tool that
-    /// can say what it wants next says it here, and the draft prints it by
-    /// the cursor.
-    fn placement_hint(&self, _placed: usize) -> Option<&'static str> {
-        None
-    }
-    /// Where the anchor the trader is still shaping really lands, given the
-    /// anchors already down and the pointer — both in screen space.
-    ///
-    /// Default: the pointer itself, which is every tool whose anchors mean
-    /// exactly where they were dropped.
-    ///
-    /// It exists because a tool of three anchors has a *shaping* phase the
-    /// raw pointer describes badly. A drag fixes a channel's trend line and
-    /// lets go with the pointer still sitting **on** that line, and a
-    /// channel's width is measured across the line and nowhere else — so the
-    /// width the pointer implies at that instant is exactly zero. The preview
-    /// draws a corridor of no width, which is a straight line, and the click
-    /// that looks like it confirms the shape commits one: a three-anchor
-    /// object that *is* a line. A tool that knows what it is refuses to be
-    /// born degenerate, and says so here rather than leaving the host to
-    /// special-case it by id.
-    ///
-    /// The host runs the preview *and* the commit through this, so the object
-    /// a click creates is always the one that was on screen when it was
-    /// clicked.
-    ///
-    /// `constrain` is what the trader is holding — see [`Constrain`]. A tool
-    /// that has an axis worth holding to says so here; the rest ignore it.
-    ///
-    /// Rate: per frame while a draft is in flight, over a handful of anchors.
-    fn pending_anchor(
-        &self,
-        _placed: &[egui::Pos2],
-        cursor: egui::Pos2,
-        _constrain: Constrain,
-    ) -> egui::Pos2 {
-        cursor
-    }
-    /// The key that arms this tool from the chart, if it has one.
-    fn shortcut(&self) -> Option<ToolShortcut> {
-        None
-    }
-    /// Where this tool's anchors land on the bar under the pointer.
-    fn anchor_snap(&self) -> AnchorSnap {
-        AnchorSnap::Pointer
-    }
-    /// The words this object holds, when its content is words rather than
-    /// geometry. `None` for every tool but the note.
-    ///
-    /// Declaring it is what earns the on-chart editor: a tool that holds
-    /// text is placed *empty*, so the host puts the caret in the object
-    /// itself the moment it lands. That used to be answered by opening the
-    /// settings panel, which typed the note in one place and showed it in
-    /// another — the eye crossing the screen between keystrokes, and the
-    /// object under the pointer reading "Note" in grey the whole time.
-    ///
-    /// Borrowed, never cloned: this is read on the frame path while the
-    /// editor is open, and a note can be a paragraph.
-    fn inline_text<'a>(&self, _payload: &'a dyn DrawingPayload) -> Option<&'a str> {
-        None
-    }
-    /// Write the words back. Paired with [`Self::inline_text`]: a tool that
-    /// answers one answers both, and the editor is the only caller.
-    fn set_inline_text(&self, _payload: &mut dyn DrawingPayload, _text: String) {}
-    /// Whether this tool's content is words at all — asked of the *tool*,
-    /// with no object in hand, on the placement path. Answering it by
-    /// building a payload just to look at it would allocate a box to learn
-    /// something that is fixed at compile time. A test holds the two answers
-    /// together, so a tool cannot claim one and implement the other.
-    fn holds_text(&self) -> bool {
-        false
-    }
-    /// Whether this tool is placed by a held drag instead of by N clicks.
-    ///
-    /// A freehand tool answers `0` from [`Self::required_points`], because
-    /// the count is whatever the gesture gave: the host starts its draft on
-    /// the press, feeds it the path, and finishes it on the release.
-    fn freehand(&self) -> bool {
-        false
-    }
-    /// The rectangle this tool actually *paints*, given the box its anchors
-    /// span and the pane it is drawn in.
-    ///
-    /// The anchor box is the default and is right for most tools: a trend
-    /// line, a rectangle, a triangle all end where their anchors do. It is
-    /// badly wrong for the ones that do not. A fixed-range profile carries two
-    /// anchors at a single price and paints a histogram across the whole price
-    /// axis; a vertical line has one anchor and paints floor to ceiling.
-    ///
-    /// Anything that has to keep clear of an object — the settings inspector,
-    /// the context bar — asks for this, not for the anchors. Placing against
-    /// the anchors of a profile means walking around a thin horizontal sliver
-    /// and landing in the middle of the figure, which is precisely the bug
-    /// this exists to make impossible.
-    fn painted_bounds(&self, anchors: egui::Rect, _chart: egui::Rect) -> egui::Rect {
-        anchors
-    }
-    /// The colour a fresh object of this tool is born in, when the stock
-    /// blue would be the wrong answer. `None` — almost every tool — takes
-    /// [`DEFAULT_DRAWING_COLOR`].
-    ///
-    /// It exists for the tools whose colour *is* their meaning: a buy mark
-    /// that arrives blue is one the trader repaints every single time. The
-    /// trader's own saved default still wins over this, because that one was
-    /// chosen rather than assumed.
-    fn default_color(&self) -> Option<egui::Color32> {
-        None
-    }
-    /// The stroke width a fresh object of this tool is born with, when the
-    /// stock hairline would be the wrong answer. The hairline rule (§D2) is
-    /// written for *annotations*; a tool that paints a derived **series** —
-    /// one value per bar competing with candle bodies — declares its weight
-    /// here instead of asking every trader to fix it by hand.
-    fn default_width_px(&self) -> Option<f32> {
-        None
-    }
-    /// The fill alpha a fresh object opens with, when the stock value would
-    /// read as "the fill is broken" for this tool's geometry.
-    fn default_fill_alpha(&self) -> Option<u8> {
-        None
-    }
-    /// The chart's right-click menu entry that places this tool at the
-    /// clicked bar, if the tool wants one. The pane sweeps the registry —
-    /// the next series tool docks with a declaration, not a pane.rs edit.
-    fn context_menu_label(&self) -> Option<&'static str> {
-        None
-    }
-    /// The rail family this tool belongs to, if any. Consecutive registry
-    /// entries with the same family id share one rail slot.
-    fn family(&self) -> Option<ToolFamily> {
-        None
-    }
-    /// Whether the tool paints an interior that the fill controls affect.
-    fn supports_fill(&self) -> bool {
-        false
-    }
-    /// Whether this tool's anchors carry a meaningful *value*.
-    ///
-    /// Almost every tool's second coordinate means something on the axis it
-    /// was drawn against, which is what binds it to one band. A vertical line
-    /// and a date range mark instants: they belong to no band, so they are
-    /// placed as [`DrawingBand::AllBands`] and painted through every band as
-    /// one object (`docs/ux/drawing-tools-2026-08.md` §D10).
-    fn value_axis(&self) -> bool {
-        true
-    }
-    /// Whether this tool only ever means something on the candles' price
-    /// axis. The mirror of [`Self::value_axis`]'s `false`: where a time-only
-    /// tool belongs to *every* band, a price-only tool belongs to the price
-    /// band whatever band it was started over — a volume profile's rows are
-    /// prices, and one drawn against a CVD axis would be the data-honesty
-    /// failure this repo refuses.
-    fn price_band_only(&self) -> bool {
-        false
-    }
-    /// Whether the tool paints a stroke the width control affects. Almost
-    /// every tool does; a text note has glyphs and no stroke at all, and a
-    /// width slider on its Style tab would move nothing.
-    fn supports_stroke_width(&self) -> bool {
-        true
-    }
-    /// The object's own glyph size, for a tool drawn as a glyph rather than
-    /// a stroke — a text note, a trade mark. `None`, the answer for almost
-    /// every tool, means the object has no such size and the context bar
-    /// offers stroke width in that slot instead.
-    ///
-    /// The size is in screen pixels and stays in screen pixels: a note about
-    /// the chart that inflates with the zoom has quietly become a second
-    /// series.
-    fn glyph_size(&self, _payload: &dyn DrawingPayload) -> Option<GlyphSize> {
-        None
-    }
-    /// Write a glyph size back. A tool that answers [`Self::glyph_size`]
-    /// must implement this, or its own control would move nothing.
-    fn set_glyph_size(&self, _payload: &mut dyn DrawingPayload, _px: f32) {}
-    /// Fresh tool-owned state for a newly placed object.
-    fn default_payload(&self) -> Box<dyn DrawingPayload> {
-        Box::new(NoPayload)
-    }
-    /// Title of the tool-owned inspector tab, if the tool brings one.
-    fn extra_tab(&self) -> Option<&'static str> {
-        None
-    }
-    /// Draw the tool-owned inspector tab. Returns whether anything was
-    /// edited (the caller folds it into the shared undo coalescing).
-    fn draw_extra_tab(
-        &self,
-        _ui: &mut egui::Ui,
-        _drawing: &mut Drawing,
-        _host: &mut dyn PresetHost,
-    ) -> bool {
-        false
-    }
-    fn paint(
-        &self,
-        painter: &egui::Painter,
-        chart_rect: egui::Rect,
-        style: DrawingStyle,
-        points: &[egui::Pos2],
-        ctxt: &DrawContext<'_>,
-    );
-    /// The part of this object that belongs *under* the candles, if any.
-    ///
-    /// Almost every tool draws nothing here, which is why the default is a
-    /// no-op: a line, a note, a Fibonacci grid is an annotation *on* the
-    /// chart, and burying it under the price would be losing it. A volume
-    /// profile is the exception that earns the pass — its histogram is
-    /// context the price is read *against*, like the liquidity map, and drawn
-    /// over the candles it tints every body it covers.
-    ///
-    /// Called before the candles, on the same geometry the over-candles pass
-    /// gets, with no selection halo and no handles: those are affordances,
-    /// and an affordance under the price is not one.
-    fn paint_under(
-        &self,
-        _painter: &egui::Painter,
-        _chart_rect: egui::Rect,
-        _style: DrawingStyle,
-        _points: &[egui::Pos2],
-        _ctxt: &DrawContext<'_>,
-    ) {
-    }
-    /// The heights on the price axis this object *is*, for the axis to tag.
-    ///
-    /// A horizontal line does not merely cross a price, it names one: the
-    /// price is the whole object, and the trader who drew it wants to read it
-    /// off the axis without hunting for where the line meets the gutter — the
-    /// yellow tag every other platform puts there. So the tool declares its
-    /// levels and [`crate::pane::ChartPane::draw_drawing_axis_tags`] writes
-    /// them, in the object's own colour.
-    ///
-    /// Screen `y`, not price, because `points` are already projected and the
-    /// projection has exactly one owner: a tool that converted back to price
-    /// itself would be a second copy of the price scale, free to drift from
-    /// the one the axis labels are drawn with. The axis reads the price back
-    /// off the same scale it labels.
-    ///
-    /// The default is empty, and that is the answer for almost every tool: a
-    /// trend line crosses every price between its ends and names none of
-    /// them, and an axis tagged with everything drawn on the chart is an axis
-    /// nobody can read. Declaring a level is opting *in* to the gutter.
-    ///
-    /// `chart_rect` is the same rect [`DrawingToolImpl::paint`] gets, and it
-    /// is here for one reason: a tool must not declare a level it is not
-    /// drawing. A horizontal ray whose anchor projects past the right edge
-    /// paints no stroke at all, and a chip on the gutter for a line that is
-    /// not there is the axis claiming something the canvas is not.
-    fn axis_levels(&self, _chart_rect: egui::Rect, _points: &[egui::Pos2]) -> AxisLevels {
-        AxisLevels::new()
-    }
-    fn hit_test(
-        &self,
-        chart_rect: egui::Rect,
-        points: &[egui::Pos2],
-        position: egui::Pos2,
-        radius_px: f32,
-        ctxt: &DrawContext<'_>,
-    ) -> bool;
-    /// Where the trader grabs this object, in screen space. `None` — the
-    /// answer for almost every tool — means the raw anchors: the point you
-    /// clicked is the point you drag.
-    ///
-    /// A tool overrides this when its anchors are not where the gesture
-    /// belongs. A channel is the case that forced the port open: its third
-    /// anchor is a corner of the corridor, so the only way to widen one was
-    /// to find a lone dot off in the distance, and only ever that one edge.
-    /// The handle for a rail belongs at the centre of that rail, and there is
-    /// one per rail.
-    fn handles(
-        &self,
-        _chart_rect: egui::Rect,
-        _points: &[egui::Pos2],
-        _ctxt: &DrawContext<'_>,
-    ) -> Option<Handles> {
-        None
-    }
-    /// Apply a drag of handle `handle` to screen position `to`, answering the
-    /// object's new screen anchors — the tool decides which anchors a handle
-    /// moves, and a handle may well move more than one.
-    ///
-    /// `None` (the default) means the plain anchor move the host already
-    /// does. A tool that overrides [`DrawingToolImpl::handles`] must override
-    /// this too: a handle that is not an anchor has no default meaning.
-    fn drag_handle(
-        &self,
-        _chart_rect: egui::Rect,
-        _points: &[egui::Pos2],
-        _handle: usize,
-        _to: egui::Pos2,
-        _ctxt: &DrawContext<'_>,
-        _constrain: Constrain,
-    ) -> Option<Handles> {
-        None
-    }
-    #[cfg(test)]
-    fn test_geometry(&self) -> (Vec<egui::Pos2>, egui::Pos2);
-}
-
-/// A cheap, copyable reference to one registered implementation.
-#[derive(Clone, Copy)]
-pub struct DrawingTool(&'static dyn DrawingToolImpl);
-
-impl DrawingTool {
-    #[must_use]
-    pub fn id(self) -> &'static str {
-        self.0.id()
-    }
-
-    /// Look up a registered tool by its stable id — how the saved favorites
-    /// list and the env hooks name tools. `None` for an id no registered
-    /// tool carries (a stale file survives a removed tool).
-    #[must_use]
-    pub fn by_id(id: &str) -> Option<Self> {
-        DRAWING_TOOLS.into_iter().find(|tool| tool.id() == id)
-    }
-
-    /// Vector icon strokes, `&[]` when the tool paints a font glyph.
-    #[must_use]
-    pub fn icon_strokes(self) -> IconStrokes {
-        self.0.icon_strokes()
-    }
-
-    /// Anchor dots of the vector icon, `&[]` when it has none.
-    #[must_use]
-    pub fn icon_dots(self) -> IconDots {
-        self.0.icon_dots()
-    }
-
-    /// The letter set into the vector icon, `None` when it needs none.
-    #[must_use]
-    pub fn icon_letter(self) -> Option<IconLetter> {
-        self.0.icon_letter()
-    }
-
-    #[must_use]
-    pub fn name(self) -> &'static str {
-        self.0.name()
-    }
-
-    #[must_use]
-    pub fn settings_title(self) -> &'static str {
-        self.0.settings_title()
-    }
-
-    #[must_use]
-    pub fn supports_fill(self) -> bool {
-        self.0.supports_fill()
-    }
-
-    #[must_use]
-    pub fn supports_stroke_width(self) -> bool {
-        self.0.supports_stroke_width()
-    }
-
-    #[must_use]
-    pub fn anchor_snap(self) -> AnchorSnap {
-        self.0.anchor_snap()
-    }
-
-    #[must_use]
-    pub fn freehand(self) -> bool {
-        self.0.freehand()
-    }
-
-    /// See [`DrawingToolImpl::painted_bounds`].
-    #[must_use]
-    pub fn painted_bounds(self, anchors: egui::Rect, chart: egui::Rect) -> egui::Rect {
-        self.0.painted_bounds(anchors, chart)
-    }
-
-    /// See [`DrawingToolImpl::axis_levels`].
-    #[must_use]
-    pub fn axis_levels(self, chart_rect: egui::Rect, points: &[egui::Pos2]) -> AxisLevels {
-        self.0.axis_levels(chart_rect, points)
-    }
-
-    /// See [`DrawingToolImpl::inline_text`].
-    #[must_use]
-    pub fn inline_text(self, payload: &dyn DrawingPayload) -> Option<&str> {
-        self.0.inline_text(payload)
-    }
-
-    /// See [`DrawingToolImpl::set_inline_text`].
-    pub fn set_inline_text(self, payload: &mut dyn DrawingPayload, text: String) {
-        self.0.set_inline_text(payload, text);
-    }
-
-    /// Whether this tool's content is words — the question the host asks a
-    /// freshly placed object to decide whether it needs the caret.
-    #[must_use]
-    pub fn holds_text(self) -> bool {
-        self.0.holds_text()
-    }
-
-    /// The stock look of a fresh object of this tool, before the trader's
-    /// own saved default is consulted.
-    #[must_use]
-    pub fn default_style(self) -> DrawingStyle {
-        let stock = DrawingStyle::default();
-        DrawingStyle {
-            color: self.0.default_color().unwrap_or(DEFAULT_DRAWING_COLOR),
-            width_px: self.0.default_width_px().unwrap_or(stock.width_px),
-            fill_alpha: self.0.default_fill_alpha().unwrap_or(stock.fill_alpha),
-        }
-    }
-
-    #[must_use]
-    pub fn context_menu_label(self) -> Option<&'static str> {
-        self.0.context_menu_label()
-    }
-
-    #[must_use]
-    pub fn glyph_size(self, drawing: &Drawing) -> Option<GlyphSize> {
-        self.0.glyph_size(drawing.payload.as_ref())
-    }
-
-    pub fn set_glyph_size(self, drawing: &mut Drawing, px: f32) {
-        self.0.set_glyph_size(drawing.payload.as_mut(), px);
-    }
-
-    #[must_use]
-    pub fn value_axis(self) -> bool {
-        self.0.value_axis()
-    }
-
-    /// The band a fresh object of this tool is placed on when the pointer is
-    /// over `band`. A time-only tool ignores the band it was drawn in — it
-    /// crosses all of them, and a band picker on it would be a control with
-    /// one correct setting. A price-only tool ignores it the other way: its
-    /// values are prices, so it lands on the price band wherever it started.
-    #[must_use]
-    pub fn band_for(self, band: &DrawingBand) -> DrawingBand {
-        if self.0.price_band_only() {
-            DrawingBand::Price
-        } else if self.value_axis() {
-            band.clone()
-        } else {
-            DrawingBand::AllBands
-        }
-    }
-
-    #[must_use]
-    pub fn default_payload(self) -> Box<dyn DrawingPayload> {
-        self.0.default_payload()
-    }
-
-    #[must_use]
-    pub fn shortcut(self) -> Option<ToolShortcut> {
-        self.0.shortcut()
-    }
-
-    #[must_use]
-    pub fn family(self) -> Option<ToolFamily> {
-        self.0.family()
-    }
-
-    #[must_use]
-    pub fn extra_tab(self) -> Option<&'static str> {
-        self.0.extra_tab()
-    }
-
-    /// Draw the tool-owned inspector tab; returns whether anything changed.
-    pub fn draw_extra_tab(
-        self,
-        ui: &mut egui::Ui,
-        drawing: &mut Drawing,
-        host: &mut dyn PresetHost,
-    ) -> bool {
-        self.0.draw_extra_tab(ui, drawing, host)
-    }
-
-    #[must_use]
-    pub fn icon(self) -> &'static str {
-        self.0.icon()
-    }
-
-    #[must_use]
-    pub fn hover_text(self) -> &'static str {
-        self.0.hover_text()
-    }
-
-    #[must_use]
-    pub fn required_points(self) -> usize {
-        self.0.required_points()
-    }
-
-    #[must_use]
-    pub fn placement_hint(self, placed: usize) -> Option<&'static str> {
-        self.0.placement_hint(placed)
-    }
-
-    /// Where the anchor under the pointer really lands while the object is
-    /// still being shaped — see [`DrawingToolImpl::pending_anchor`].
-    #[must_use]
-    pub fn pending_anchor(
-        self,
-        placed: &[egui::Pos2],
-        cursor: egui::Pos2,
-        constrain: Constrain,
-    ) -> egui::Pos2 {
-        self.0.pending_anchor(placed, cursor, constrain)
-    }
-
-    /// Paint the part of the object that goes under the candles — see
-    /// [`DrawingToolImpl::paint_under`]. No halo, no handles.
-    pub fn paint_under(
-        self,
-        painter: &egui::Painter,
-        chart_rect: egui::Rect,
-        style: DrawingStyle,
-        points: &[egui::Pos2],
-        ctxt: &DrawContext<'_>,
-    ) {
-        self.0.paint_under(painter, chart_rect, style, points, ctxt);
-    }
-
-    /// Paint the object. Selection adds a halo *under* the geometry and, when
-    /// `show_handles` (not locked), white anchor handles on top — the object's
-    /// configured colour keeps carrying meaning either way.
-    pub fn paint(
-        self,
-        painter: &egui::Painter,
-        chart_rect: egui::Rect,
-        style: DrawingStyle,
-        points: &[egui::Pos2],
-        ctxt: &DrawContext<'_>,
-        show_handles: bool,
-    ) {
-        if ctxt.selected {
-            let halo_style = DrawingStyle {
-                color: SELECTION_HALO_COLOR,
-                width_px: style.width_px + SELECTION_HALO_EXTRA_WIDTH_PX,
-                fill_alpha: 0,
-            };
-            let halo_ctxt = DrawContext {
-                halo: true,
-                content_editing: false,
-                ..*ctxt
-            };
-            self.0
-                .paint(painter, chart_rect, halo_style, points, &halo_ctxt);
-        }
-        self.0.paint(painter, chart_rect, style, points, ctxt);
-        if ctxt.selected && show_handles {
-            let ring = egui::Stroke::new(SELECTED_ANCHOR_RING_WIDTH_PX, theme::ACCENT);
-            for point in self.handles(chart_rect, points, ctxt) {
-                painter.circle_filled(point, SELECTED_ANCHOR_RADIUS_PX, SELECTED_ANCHOR_FILL);
-                painter.circle_stroke(point, SELECTED_ANCHOR_RADIUS_PX, ring);
-            }
-        }
-    }
-
-    /// The grab points of this object — the tool's own when it declares them,
-    /// the raw anchors otherwise. Paint, hit-test and drag all ask here, so
-    /// what the trader sees is exactly what they can grab.
-    #[must_use]
-    pub fn handles(
-        self,
-        chart_rect: egui::Rect,
-        points: &[egui::Pos2],
-        ctxt: &DrawContext<'_>,
-    ) -> Handles {
-        self.0
-            .handles(chart_rect, points, ctxt)
-            .unwrap_or_else(|| points.iter().copied().collect())
-    }
-
-    /// Whether this tool's handles *are* its anchors — true for almost every
-    /// tool. A host that can only express "move anchor N" (the cross-pane
-    /// shared edit) asks this before offering a handle at all, rather than
-    /// leaving a grab point that moves something other than what it sits on.
-    #[must_use]
-    pub fn handles_are_anchors(
-        self,
-        chart_rect: egui::Rect,
-        points: &[egui::Pos2],
-        ctxt: &DrawContext<'_>,
-    ) -> bool {
-        self.0.handles(chart_rect, points, ctxt).is_none()
-    }
-
-    /// New screen anchors after dragging `handle` to `to`, when the tool owns
-    /// the gesture. `None` means the host's plain "this handle is anchor
-    /// `handle`" move.
-    #[must_use]
-    pub fn drag_handle(
-        self,
-        chart_rect: egui::Rect,
-        points: &[egui::Pos2],
-        handle: usize,
-        to: egui::Pos2,
-        ctxt: &DrawContext<'_>,
-        constrain: Constrain,
-    ) -> Option<Handles> {
-        self.0
-            .drag_handle(chart_rect, points, handle, to, ctxt, constrain)
-    }
-
-    #[must_use]
-    pub fn hit_test(
-        self,
-        chart_rect: egui::Rect,
-        points: &[egui::Pos2],
-        position: egui::Pos2,
-        radius_px: f32,
-        ctxt: &DrawContext<'_>,
-    ) -> bool {
-        self.handles(chart_rect, points, ctxt)
-            .iter()
-            .any(|point| point.distance_sq(position) <= radius_px * radius_px)
-            || self
-                .0
-                .hit_test(chart_rect, points, position, radius_px, ctxt)
-    }
-
-    #[cfg(test)]
-    fn test_geometry(self) -> (Vec<egui::Pos2>, egui::Pos2) {
-        self.0.test_geometry()
-    }
-}
-
-impl PartialEq for DrawingTool {
-    fn eq(&self, other: &Self) -> bool {
-        self.id() == other.id()
-    }
-}
-
-impl Eq for DrawingTool {}
-
-impl fmt::Debug for DrawingTool {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        formatter
-            .debug_tuple("DrawingTool")
-            .field(&self.id())
-            .finish()
-    }
 }
 
 macro_rules! register_drawing_tools {
@@ -1498,763 +837,6 @@ pub struct Drawings {
     /// Snapshot taken when a pointer gesture starts; committed (as one undo
     /// entry) on release, so a whole drag coalesces into one command.
     gesture_baseline: Option<UndoEntry>,
-}
-
-impl Drawings {
-    /// How many times the collection has changed. See [`Self::revision`]'s
-    /// field: equal readings mean nothing to write.
-    #[must_use]
-    pub fn revision(&self) -> u64 {
-        self.revision
-    }
-
-    /// Empty the store and hand back what it held — the way a pane puts its
-    /// drawings away when the layout or the market under it changes.
-    ///
-    /// The undo history goes with them: it described objects that are no
-    /// longer on this chart, and an undo that brought back a mark from
-    /// another market would be the very confusion the scoping exists to
-    /// prevent. Ids are not rewound.
-    pub fn take_all(&mut self) -> Vec<Drawing> {
-        self.revision += 1;
-        self.draft = None;
-        self.selected = None;
-        self.undo.clear();
-        self.redo.clear();
-        self.gesture_baseline = None;
-        std::mem::take(&mut self.items)
-    }
-
-    /// Fill an emptied store from saved objects.
-    ///
-    /// An object keeps the id it arrives with when that id is free here, so
-    /// whatever named it — a strategy armed on it, an agent's annotation —
-    /// still does; the counter moves past it so nothing later collides. An
-    /// object with no id (`0`), or one whose id this store already gave out,
-    /// takes a fresh one.
-    ///
-    /// Anchors keep the bar offsets they came with; the pane re-anchors them
-    /// against its own series right after, exactly as after a re-cut.
-    pub fn adopt(&mut self, items: impl IntoIterator<Item = Drawing>) {
-        self.revision += 1;
-        for mut drawing in items {
-            let wanted = drawing.id.0;
-            let taken = self.items.iter().any(|held| held.id.0 == wanted);
-            if wanted == 0 || taken {
-                drawing.id = self.alloc_id();
-            } else {
-                self.next_id = self.next_id.max(wanted);
-            }
-            self.items.push(drawing);
-        }
-    }
-
-    #[must_use]
-    pub fn items(&self) -> &[Drawing] {
-        &self.items
-    }
-
-    /// Mutable access for **derived-state refresh only** — `frvp::refresh`
-    /// bringing cached profiles up to date. Never a user edit: nothing
-    /// reached through here may participate in payload equality, or a
-    /// refresh would register as an edit against the undo snapshots
-    /// (`Self::record` compares them). The cache exclusion in
-    /// `FrvpPayload::eq` is the other half of this contract.
-    #[must_use]
-    pub(crate) fn items_mut(&mut self) -> &mut [Drawing] {
-        &mut self.items
-    }
-
-    /// The in-flight draft, under the same derived-state-only contract as
-    /// [`Self::items_mut`] — the profile refresh folds it so the histogram
-    /// is live while the range is still being placed.
-    #[must_use]
-    pub(crate) fn draft_mut(&mut self) -> Option<&mut Drawing> {
-        self.draft.as_mut()
-    }
-
-    /// How many objects are painted on the tab's other panes as well — the
-    /// per-frame reprojection cost, in the health summary.
-    #[must_use]
-    pub fn shared_count(&self) -> usize {
-        self.items.iter().filter(|item| item.shared()).count()
-    }
-
-    fn alloc_id(&mut self) -> DrawingId {
-        self.next_id += 1;
-        DrawingId(self.next_id)
-    }
-
-    /// Where the object with this identity currently sits, if it still
-    /// exists. The answer is only good for this frame: every reorder or
-    /// delete moves it, which is the whole reason callers hold the id.
-    #[must_use]
-    pub fn index_of(&self, id: DrawingId) -> Option<usize> {
-        self.items.iter().position(|item| item.id == id)
-    }
-
-    /// Rename the object at `index` as one undo step. Whitespace-only input
-    /// clears the name back to the derived label.
-    pub fn rename_at(&mut self, index: usize, name: &str) {
-        if index >= self.items.len() {
-            return;
-        }
-        let trimmed = name.trim();
-        let name = (!trimmed.is_empty()).then(|| trimmed.to_owned());
-        let before = self.snapshot();
-        self.items[index].name = name;
-        self.record(before);
-    }
-
-    #[must_use]
-    pub fn selected(&self) -> Option<usize> {
-        self.selected
-    }
-
-    pub fn select(&mut self, selected: Option<usize>) {
-        self.selected = selected.filter(|&index| index < self.items.len());
-    }
-
-    #[must_use]
-    pub fn selected_mut(&mut self) -> Option<&mut Drawing> {
-        self.selected.and_then(|index| self.items.get_mut(index))
-    }
-
-    #[must_use]
-    pub fn draft(&self) -> Option<&Drawing> {
-        self.draft.as_ref()
-    }
-
-    #[must_use]
-    pub fn draft_len(&self) -> usize {
-        self.draft.as_ref().map_or(0, |draft| draft.points.len())
-    }
-
-    #[must_use]
-    pub fn all_hidden(&self) -> bool {
-        self.all_hidden
-    }
-
-    /// Whether the object at `index` paints and hit-tests this frame.
-    #[must_use]
-    pub fn is_visible(&self, index: usize) -> bool {
-        !self.all_hidden && self.items.get(index).is_some_and(|item| !item.hidden)
-    }
-
-    fn snapshot(&self) -> UndoEntry {
-        UndoEntry {
-            items: self.items.clone(),
-            all_hidden: self.all_hidden,
-        }
-    }
-
-    /// Push `before` as one undo step if the store actually changed since.
-    fn record(&mut self, before: UndoEntry) {
-        if before == self.snapshot() {
-            return;
-        }
-        self.revision += 1;
-        self.undo.push(before);
-        if self.undo.len() > UNDO_HISTORY_LIMIT {
-            self.undo.remove(0);
-        }
-        self.redo.clear();
-    }
-
-    /// Whether a coalescing gesture is open on this store.
-    ///
-    /// Test-only, and it exists for one question: when a pane edits a mark it
-    /// does not own, did the gesture open on the store that holds the mark?
-    /// The answer is invisible from outside otherwise — the baseline is
-    /// private, as it should be. The layout store asks it too: a drag in
-    /// flight is not a change to write yet.
-    #[must_use]
-    pub fn in_gesture(&self) -> bool {
-        self.gesture_baseline.is_some()
-    }
-
-    /// Start coalescing: the next [`Self::commit_gesture`] records everything
-    /// mutated in between as a single undo entry. Idempotent within a gesture.
-    pub fn begin_gesture(&mut self) {
-        if self.gesture_baseline.is_none() {
-            self.gesture_baseline = Some(self.snapshot());
-        }
-    }
-
-    /// End coalescing. A gesture that changed nothing records nothing.
-    pub fn commit_gesture(&mut self) {
-        if let Some(baseline) = self.gesture_baseline.take() {
-            self.record(baseline);
-        }
-    }
-
-    /// Record an already-applied edit to one object, given its pre-edit
-    /// state. Used by the inspector to coalesce slider/color gestures.
-    pub fn record_edit_of(&mut self, index: usize, before_drawing: Drawing) {
-        let mut before = self.snapshot();
-        let Some(slot) = before.items.get_mut(index) else {
-            return;
-        };
-        *slot = before_drawing;
-        self.record(before);
-    }
-
-    #[cfg(test)]
-    pub(crate) fn undo_depth(&self) -> usize {
-        self.undo.len()
-    }
-
-    fn restore(&mut self, entry: UndoEntry) {
-        self.revision += 1;
-        self.items = entry.items;
-        self.all_hidden = entry.all_hidden;
-        self.draft = None;
-        self.selected = self.selected.filter(|&index| index < self.items.len());
-    }
-
-    pub fn undo(&mut self) -> bool {
-        let Some(entry) = self.undo.pop() else {
-            return false;
-        };
-        self.redo.push(self.snapshot());
-        self.restore(entry);
-        true
-    }
-
-    pub fn redo(&mut self) -> bool {
-        let Some(entry) = self.redo.pop() else {
-            return false;
-        };
-        self.undo.push(self.snapshot());
-        self.restore(entry);
-        true
-    }
-
-    /// [`Self::place_with`] with the tool's stock look — the test-side
-    /// shorthand; the app always goes through `place_with` to honour the
-    /// trader's saved defaults.
-    #[cfg(test)]
-    pub fn place(&mut self, tool: DrawingTool, point: ChartPoint) -> bool {
-        self.place_on(tool, &DrawingBand::Price, point)
-    }
-
-    /// [`Self::place`] on a named band.
-    #[cfg(test)]
-    pub fn place_on(&mut self, tool: DrawingTool, band: &DrawingBand, point: ChartPoint) -> bool {
-        self.place_with(tool, band, point, |tool| NewDrawing {
-            style: DrawingStyle::default(),
-            payload: tool.default_payload(),
-        })
-    }
-
-    /// [`Self::place`] with a caller-chosen look for a new draft — how the
-    /// app applies the trader's saved defaults to newly created objects only.
-    /// Objects that already exist are never touched by a default changing.
-    pub fn place_with(
-        &mut self,
-        tool: DrawingTool,
-        band: &DrawingBand,
-        point: ChartPoint,
-        new_drawing: impl FnOnce(DrawingTool) -> NewDrawing,
-    ) -> bool {
-        if self.draft.as_ref().is_none_or(|draft| draft.tool != tool) {
-            let fresh = new_drawing(tool);
-            let id = self.alloc_id();
-            self.draft = Some(Drawing {
-                id,
-                author: None,
-                name: None,
-                tool,
-                points: Vec::with_capacity(tool.required_points()),
-                // The band the *first* anchor landed in owns the whole
-                // object: an in-flight draft that changed axis halfway
-                // through would have anchors in two value spaces.
-                band: tool.band_for(band),
-                style: fresh.style,
-                locked: false,
-                hidden: false,
-                scope: DrawingScope::default(),
-                foreign_market: false,
-                off_series: false,
-                payload: fresh.payload,
-            });
-        }
-        let draft = self.draft.as_mut().expect("draft was installed above");
-        draft.points.push(point);
-        // A freehand tool declares no anchor count: its draft is finished by
-        // the release, through `finish_draft`, never by arithmetic here.
-        if tool.required_points() > 0 && draft.points.len() == tool.required_points() {
-            let before = self.snapshot();
-            self.items
-                .push(self.draft.take().expect("draft has points"));
-            self.selected = Some(self.items.len() - 1);
-            // Drawing while hide-all is engaged releases it (audit M8): the
-            // act of placing a new object is the strongest possible request
-            // to see drawings, and a mark that vanishes on the click that
-            // finished it reads as a broken tool. One undo entry with the
-            // placement — undoing the object restores the hidden state too.
-            self.all_hidden = false;
-            self.record(before);
-            true
-        } else {
-            false
-        }
-    }
-
-    pub fn cancel_draft(&mut self) {
-        self.draft = None;
-    }
-
-    /// Finish a freehand draft: what the release does for a tool whose
-    /// anchor count is whatever the hand gave.
-    ///
-    /// A stroke of fewer than two points is a click that missed, not a
-    /// drawing — it is dropped rather than stored as an invisible object the
-    /// trader can neither see nor select to delete.
-    pub fn finish_draft(&mut self) -> bool {
-        let Some(draft) = self.draft.take() else {
-            return false;
-        };
-        if draft.points.len() < 2 {
-            return false;
-        }
-        let before = self.snapshot();
-        self.items.push(draft);
-        self.selected = Some(self.items.len() - 1);
-        // Same rule as a clicked placement: drawing releases hide-all, and
-        // it does so inside the one undo entry the gesture records.
-        self.all_hidden = false;
-        self.record(before);
-        true
-    }
-
-    /// Backspace during placement: drop the last placed anchor; dropping the
-    /// only one cancels the draft.
-    pub fn remove_last_draft_anchor(&mut self) {
-        if let Some(draft) = &mut self.draft {
-            draft.points.pop();
-            if draft.points.is_empty() {
-                self.draft = None;
-            }
-        }
-    }
-
-    /// Duplicate the selected object as one undo entry: the copy lands
-    /// `offset_bars` to the right, unlocked, and becomes the selection.
-    ///
-    /// Returns the pair so the layer above can carry whatever rides the
-    /// drawing without living in it. `#[must_use]`: dropping the answer is
-    /// how a copied band silently loses its bot.
-    #[must_use]
-    pub fn duplicate_selected(&mut self, offset_bars: f32) -> Option<Duplicated> {
-        let index = self.selected.filter(|&index| index < self.items.len())?;
-        let before = self.snapshot();
-        let mut copy = self.items[index].clone();
-        // A copy is a new object: its own identity, and never the
-        // original's name — two drawings answering to "congestão 108k"
-        // would make every reference ambiguous.
-        copy.id = self.alloc_id();
-        copy.name = None;
-        for point in &mut copy.points {
-            point.bar += offset_bars;
-            // The offset moves the copy in *bar* space, and the instant an
-            // anchor remembers is what survives a re-cut: `reanchor` puts
-            // every timestamped anchor back where its own market moment
-            // landed. Carrying the source's instants would therefore snap
-            // the copy onto the original at the next bar-spec change,
-            // replay seek, reconnect or symbol switch — two rectangles at
-            // one place, and since this branch hangs a bot on the copy, two
-            // bots there too. The copy is a mark the trader placed just
-            // now, at no market moment of its own; that is what an anchor
-            // dropped past the newest bar already carries, and it is the
-            // honest reading here.
-            point.time_ms = None;
-        }
-        copy.locked = false;
-        let duplicated = Duplicated {
-            source: self.items[index].id,
-            copy: copy.id,
-        };
-        self.items.push(copy);
-        self.selected = Some(self.items.len() - 1);
-        self.record(before);
-        Some(duplicated)
-    }
-
-    /// Re-express every anchor against a series that was cut again — a
-    /// timeframe or bar-kind switch, a replay seek, a reconnect, a symbol
-    /// change.
-    ///
-    /// A bar index means nothing across two cuts of the tape; the market
-    /// instant each anchor captured at placement does, and it is the same
-    /// coordinate that already carries a drawing between the panes of a tab
-    /// (`docs/ux/drawing-tools-2026-08.md` §D7). So the object is not
-    /// discarded and not left pointing at a stale index: it is asked where
-    /// its own timestamps landed.
-    ///
-    /// `slot_of` answers where a market instant sits on the new series, or
-    /// `None` when the series does not reach it — before its first bar, or on
-    /// an instrument that never traded at that moment. Those anchors clamp to
-    /// the nearest edge and the object is flagged [`Drawing::off_series`], so
-    /// it fades and says so rather than pretending to sit on data.
-    ///
-    /// An anchor with no timestamp at all is one dropped past the newest bar,
-    /// where the tape has written nothing (see `ChartPoint::time_ms`). It has
-    /// no instant to look up, so it keeps its distance past the end of the
-    /// series instead — which is exactly where the trader put it.
-    ///
-    /// Rate: once per re-cut, never per frame or per trade, over a handful of
-    /// objects. The undo stacks travel with the live items for the same
-    /// reason [`Self::shift_bars`] moves them — undoing later must not
-    /// resurrect coordinates from a series that no longer exists.
-    pub fn reanchor(
-        &mut self,
-        old_slots: usize,
-        new_slots: usize,
-        slot_of: impl Fn(i64) -> Option<f32>,
-    ) {
-        #[allow(clippy::cast_precision_loss)]
-        let past_end = new_slots as f32 - old_slots as f32;
-        let reanchor_all = |items: &mut [Drawing]| {
-            for drawing in items {
-                let mut off_series = false;
-                for point in &mut drawing.points {
-                    let Some(time) = point.time_ms else {
-                        point.bar += past_end;
-                        continue;
-                    };
-                    match slot_of(time) {
-                        Some(slot) => point.bar = slot,
-                        None => {
-                            off_series = true;
-                            point.bar = 0.0;
-                        }
-                    }
-                }
-                drawing.off_series = off_series;
-            }
-        };
-        reanchor_all(&mut self.items);
-        if let Some(draft) = self.draft.as_mut() {
-            reanchor_all(std::slice::from_mut(draft));
-        }
-        for entry in self.undo.iter_mut().chain(self.redo.iter_mut()) {
-            reanchor_all(&mut entry.items);
-        }
-        if let Some(baseline) = &mut self.gesture_baseline {
-            reanchor_all(&mut baseline.items);
-        }
-    }
-
-    /// Mark every object as belonging to a market this tab no longer shows.
-    ///
-    /// Called on the one transition that changes the instrument under the
-    /// marks. Everything present at that moment was drawn on the old market;
-    /// anything placed afterwards is on the new one and starts clean.
-    ///
-    /// The undo stacks travel with the live items, for the same reason
-    /// [`Self::reanchor`] moves them: undoing back to a state from before the
-    /// switch must not restore a mark that claims to be a level on this
-    /// instrument.
-    pub fn mark_market_changed(&mut self) {
-        let mark = |items: &mut [Drawing]| {
-            for drawing in items {
-                drawing.foreign_market = true;
-            }
-        };
-        mark(&mut self.items);
-        if let Some(draft) = self.draft.as_mut() {
-            mark(std::slice::from_mut(draft));
-        }
-        for entry in self.undo.iter_mut().chain(self.redo.iter_mut()) {
-            mark(&mut entry.items);
-        }
-        if let Some(baseline) = &mut self.gesture_baseline {
-            mark(&mut baseline.items);
-        }
-    }
-
-    /// Rewrite the market instants behind one object's anchors, after a move
-    /// that changed their bar positions.
-    ///
-    /// [`Self::translate_selected`] and the keyboard nudge shift bar indices
-    /// directly; the timestamp behind each anchor is what every *other* pane
-    /// reads, so leaving it stale would drag a mark on one chart and leave
-    /// its shared twin standing where it used to be. Only the pane knows how
-    /// to name the instant under a slot, so it hands the answers back here.
-    pub fn set_times(&mut self, index: usize, times: &[Option<i64>]) {
-        let Some(drawing) = self.items.get_mut(index) else {
-            return;
-        };
-        for (point, time) in drawing.points.iter_mut().zip(times) {
-            point.time_ms = *time;
-        }
-    }
-
-    // There is deliberately no `clear`. A re-cut of the bars used to wipe the
-    // store, on the reasoning that a bar index cannot survive one; the anchors
-    // carry market time, so they are re-expressed instead
-    // ([`Self::reanchor`]). The only way a drawing leaves is the trader
-    // removing it — [`Self::delete_selected`] or [`Self::delete_all`], both of
-    // which are undoable.
-
-    pub fn shift_bars(&mut self, delta: isize) {
-        if delta == 0 {
-            return;
-        }
-        let delta = delta as f32;
-        let shift = |items: &mut Vec<Drawing>| {
-            for drawing in items {
-                for point in &mut drawing.points {
-                    point.bar += delta;
-                }
-            }
-        };
-        shift(&mut self.items);
-        if let Some(draft) = &mut self.draft {
-            for point in &mut draft.points {
-                point.bar += delta;
-            }
-        }
-        // History snapshots hold the same bar-index coordinates, so a prepend
-        // shifts them too — undoing later must not re-anchor objects to bars
-        // that moved underneath them.
-        for entry in self.undo.iter_mut().chain(self.redo.iter_mut()) {
-            shift(&mut entry.items);
-        }
-        if let Some(baseline) = &mut self.gesture_baseline {
-            shift(&mut baseline.items);
-        }
-    }
-
-    /// Remove every drawing as one undoable step, and report how many went.
-    /// Locked objects go too: the caller gates this behind a count-bearing
-    /// confirmation, and a lock protects against a stray click, not against
-    /// an explicit clear-everything — `Ctrl+Z` brings the whole set back.
-    pub fn delete_all(&mut self) -> usize {
-        let count = self.items.len();
-        if count == 0 {
-            return 0;
-        }
-        let before = self.snapshot();
-        self.items.clear();
-        self.selected = None;
-        self.record(before);
-        count
-    }
-
-    /// One delete command for every trigger (button, manager, keyboard).
-    /// A locked object is never deleted without `force`.
-    pub fn delete_selected(&mut self, force: bool) -> DeleteOutcome {
-        let Some(index) = self.selected.filter(|&index| index < self.items.len()) else {
-            return DeleteOutcome::NothingSelected;
-        };
-        if self.items[index].locked && !force {
-            return DeleteOutcome::NeedsConfirmation;
-        }
-        let before = self.snapshot();
-        self.items.remove(index);
-        self.selected = None;
-        self.record(before);
-        DeleteOutcome::Deleted
-    }
-
-    pub fn set_selected_locked(&mut self, locked: bool) {
-        if let Some(index) = self.selected {
-            self.set_locked_at(index, locked);
-        }
-    }
-
-    pub fn set_selected_hidden(&mut self, hidden: bool) {
-        if let Some(index) = self.selected {
-            self.set_hidden_at(index, hidden);
-        }
-    }
-
-    /// Remove one object by identity, wherever it sits and whatever is
-    /// selected — the removal an operator's own annotation gets, and the one
-    /// the trader's "remove what an assistant placed" gesture repeats.
-    /// Locked objects placed by an operator are included: a lock the trader
-    /// did not set is not a reason to keep someone else's mark.
-    pub fn remove_by_id(&mut self, id: DrawingId) -> bool {
-        let Some(index) = self.items.iter().position(|drawing| drawing.id == id) else {
-            return false;
-        };
-        let before = self.snapshot();
-        self.items.remove(index);
-        self.selected = match self.selected {
-            Some(selected) if selected == index => None,
-            Some(selected) if selected > index => Some(selected - 1),
-            other => other,
-        };
-        self.record(before);
-        true
-    }
-
-    /// Remove every object an operator other than the trader placed, and
-    /// report how many went. One gesture, one undo entry — the trader's way
-    /// back from an assistant that drew too much.
-    pub fn remove_authored(&mut self) -> usize {
-        if !self.items.iter().any(|drawing| drawing.author.is_some()) {
-            return 0;
-        }
-        let before = self.snapshot();
-        let previous = self.items.len();
-        let selected_id = self
-            .selected
-            .and_then(|index| self.items.get(index))
-            .map(|drawing| drawing.id);
-        self.items.retain(|drawing| drawing.author.is_none());
-        self.selected =
-            selected_id.and_then(|id| self.items.iter().position(|drawing| drawing.id == id));
-        self.record(before);
-        previous - self.items.len()
-    }
-
-    /// How many objects on this pane were placed by an operator other than
-    /// the trader — what the sweep gesture shows before it offers itself.
-    #[must_use]
-    pub fn authored_count(&self) -> usize {
-        self.items
-            .iter()
-            .filter(|drawing| drawing.author.is_some())
-            .count()
-    }
-
-    /// Say who placed one object, after the fact.
-    ///
-    /// The annotate tier stamps authorship as it places, so this exists for
-    /// the tests that need an object to *be* an assistant's without standing
-    /// up a gateway to produce one.
-    #[cfg(test)]
-    pub fn set_author_at(&mut self, index: usize, author: Option<DrawingAuthor>) {
-        if let Some(drawing) = self.items.get_mut(index) {
-            drawing.author = author;
-        }
-    }
-
-    pub fn set_locked_at(&mut self, index: usize, locked: bool) {
-        let before = self.snapshot();
-        if let Some(drawing) = self.items.get_mut(index) {
-            drawing.locked = locked;
-            self.record(before);
-        }
-    }
-
-    pub fn set_hidden_at(&mut self, index: usize, hidden: bool) {
-        let before = self.snapshot();
-        if let Some(drawing) = self.items.get_mut(index) {
-            drawing.hidden = hidden;
-            self.record(before);
-        }
-    }
-
-    /// Z-order: painting walks the list front-to-back, hit-testing walks it
-    /// back-to-front, so the last item is the topmost object.
-    pub fn bring_to_front(&mut self, index: usize) {
-        if index >= self.items.len() || index + 1 == self.items.len() {
-            return;
-        }
-        let before = self.snapshot();
-        let drawing = self.items.remove(index);
-        self.items.push(drawing);
-        // Selection follows the object, not the slot it used to occupy.
-        self.selected = self.selected.map(|selected| {
-            if selected == index {
-                self.items.len() - 1
-            } else if selected > index {
-                selected - 1
-            } else {
-                selected
-            }
-        });
-        self.record(before);
-    }
-
-    /// The *opening* hidden state, seeded as a pane is built.
-    ///
-    /// Deliberately not [`Self::set_all_hidden`]: that one records an undo
-    /// entry, which is right for a click and wrong for a pane that has just
-    /// come into existence. A time pane seeded through the recording setter
-    /// opens with a non-empty history on a store holding zero objects, so the
-    /// trader's first Ctrl+Z there un-hides drawings instead of doing nothing.
-    pub fn open_all_hidden(&mut self, hidden: bool) {
-        self.all_hidden = hidden;
-    }
-
-    pub fn set_all_hidden(&mut self, hidden: bool) {
-        let before = self.snapshot();
-        self.all_hidden = hidden;
-        self.record(before);
-    }
-
-    /// Whether every drawing is individually locked (used by the toolbox's
-    /// lock-all toggle). An empty collection is not "all locked".
-    #[must_use]
-    pub fn all_locked(&self) -> bool {
-        !self.items.is_empty() && self.items.iter().all(|item| item.locked)
-    }
-
-    /// Reversible bulk protection: locks (or unlocks) every drawing as one
-    /// undo entry. Never deletes anything.
-    pub fn set_all_locked(&mut self, locked: bool) {
-        let before = self.snapshot();
-        for item in &mut self.items {
-            item.locked = locked;
-        }
-        self.record(before);
-    }
-
-    /// Rigid translation of the selected object. Locked geometry stays put.
-    pub fn translate_selected(&mut self, delta_bar: f32, delta_price: f64) {
-        let Some(drawing) = self.selected_mut() else {
-            return;
-        };
-        if drawing.locked {
-            return;
-        }
-        for point in &mut drawing.points {
-            point.bar += delta_bar;
-            point.price += delta_price;
-        }
-    }
-
-    /// Move one anchor of one object. Locked geometry stays put.
-    pub fn move_anchor(
-        &mut self,
-        drawing_index: usize,
-        point_index: usize,
-        point: ChartPoint,
-    ) -> bool {
-        let Some(drawing) = self.items.get_mut(drawing_index) else {
-            return false;
-        };
-        if drawing.locked {
-            return false;
-        }
-        let Some(anchor) = drawing.points.get_mut(point_index) else {
-            return false;
-        };
-        *anchor = point;
-        true
-    }
-
-    /// Replace every anchor of one object at once — what a tool-owned handle
-    /// drag produces, because a handle that moves a rail moves the anchors
-    /// that define it together. Locked geometry stays put, and an anchor
-    /// count that does not match is refused rather than reshaping the object
-    /// into something the tool cannot paint.
-    pub fn set_points(&mut self, drawing_index: usize, points: &[ChartPoint]) -> bool {
-        let Some(drawing) = self.items.get_mut(drawing_index) else {
-            return false;
-        };
-        if drawing.locked || drawing.points.len() != points.len() {
-            return false;
-        }
-        drawing.points.clear();
-        drawing.points.extend_from_slice(points);
-        true
-    }
 }
 
 pub(super) fn drawing_stroke(style: DrawingStyle) -> egui::Stroke {
